scpr-memcached 0.1.0: First checkin for memcache cookbook
Offers an `scpr_memcached` resource, which can be used to create one or more
memcached instances on a server. You can also specify a hash under
`node.scpr_memcached.instances`. In either case, we expect a unique key,
`memory` as number of megabytes, and `port` as an available port number.